pyproject.toml
[project]
name = "fastapi-python-reacttalon-postgres"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
